Fix formatting issues

+++ src/markov_matrix.c
@@ -582,7 +582,7 @@ int mm_sample_state(MarkovMatrix *M, int state) {
 
 /* as above but by character */
 char mm_sample_backgd(char *labels, Vector *backgd) {
-  if (strlen(labels) != backgd->size) die("mm_sample_backgd: got num_labels=%i but backgd->size=%i\n", strlen(labels), backgd->size);
+  if (strlen(labels) != backgd->size) die("mm_sample_backgd: got num_labels=%i but backgd->size=%i\n", (int)strlen(labels), backgd->size);
   return labels[pv_draw_idx(backgd)];
 }
